Create an embedding model using sentence-transformers

Initially this function is built upon the sentence_transformers Python library, but it may be expanded to accept other frameworks. You should feed in your documents as a list. You can use hardware accelerators e.g. GPUs, to speed up computation.

Usage

bt_make_embedder_st(model)

Arguments

model

Name of embedding model as a string (not case sensitive)

Value

an embedding model, formed according to the model defined, that can be input to be_do_embedding to create document embeddings

Examples

embedder <- bt_make_embedder_st("all-miniLM-L6-v2")
#> Error in py_call_impl(callable, call_args$unnamed, call_args$named): ImportError: Can't connect to HTTPS URL because the SSL module is not available.
#> Run `reticulate::py_last_error()` for details.

embedder <- bt_make_embedder_st("aLL-minilm-l6-v2")
#> Error in py_call_impl(callable, call_args$unnamed, call_args$named): ImportError: Can't connect to HTTPS URL because the SSL module is not available.
#> Run `reticulate::py_last_error()` for details.