fit a bertopic model to cleaned data

fit a bertopic model to cleaned data

Usage

fit_transform_model(
cleaned_text,
calculated_embeddings = NULL,
min_topic_size = 10,
nr_topics = NULL,
ngram_range = tuple(1,1),
embedding_model = "all-MiniLM-L6-v2",
accelerator = "mps",
diversity = 0.1,
stopwords = TRUE,
random_state = NULL)

Arguments

cleaned_text

cleaned text column that the model should be fit with

calculated_embeddings

if embeddings are already calculated input embeddings as matrix

min_topic_size

minimum topic size

nr_topics

the number of topics to find within the dataset

ngram_range

ngram range for topic representation - input must be of type tuple: use the reticulate function reticulate::tuple()

embedding_model

which embedding model to use or was used to produce embeddings

accelerator

accelerator to use - default is mps, use NULL if none

diversity

diversity of topic representation (1 = diverse, 0 = not diverse)

stopwords

whether or not to remove stopwords in topic representations

random_state

random state to pass to umap

Value

list containing fitted bertopic model and embeddings used to fit the model