create top terms charts for topic modelling completed using bertopic

create top terms charts for topic modelling completed using bertopic

Usage

viz_top_terms(
merged_df,
text_var = text_clean,
 topic_var = topic,
stopwords = TRUE,
hashtags = TRUE,
mentions = TRUE,
top_n = 15,
n_row = 2,
min_freq = 25,
include_outliers = FALSE,
type = "lollipops")

Arguments

merged_df

df output from makedf function.Can be any df that includes a topic column

text_var

text top terms to be extracted from

topic_var

column containing topic variable

stopwords

remove (english) stopwords?

hashtags

remove hashtags?

mentions

remove mentions?

top_n

number of terms to extract

n_row

number of rows the plots should take up

min_freq

minimum number of times a term should appear for it to be considered

include_outliers

include the -1 (outlier) bertopic category

type

lollipop or bar chart?

Value

list of all_terms and max_only top term bar charts