Build a BERTopic model

Keep *_model = NULL to proceed with a model made from default parameters (see each individual make_* function for parameters). However, it is not advisable to accept default parameters for each model; you should tune each model according to your dataset and the business question you are answering.

bt_compile_model(
  ...,
  embedding_model = NULL,
  reduction_model = NULL,
  clustering_model = NULL,
  vectoriser_model = NULL,
  ctfidf_model = NULL
)

Arguments

...

Additional arguments sent to bertopic.BERTopic()

embedding_model

Model for creating embeddings (Python object)

reduction_model

Model for reducing embeddings' dimensions (Python object)

clustering_model

Model for clustering (Python object)

vectoriser_model

Model for vectorising input for topic representations (Python object)

ctfidf_model

Model for performing class-based tf-idf (ctf-idf) (Python object)

Value

a BERTopic model

Examples